# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Brand & Marketing OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get a social post by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a social post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePostBody"
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}/generate-image:
    post:
      operationId: generatePostImage
      tags: [posts]
      summary: Generate an AI image for a post using its imagePrompt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePostImageBody"
      responses:
        "200":
          description: Image generated and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /posts/{id}/regenerate:
    post:
      operationId: regeneratePost
      tags: [posts]
      summary: Regenerate AI content for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post regenerated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        day:
          type: integer
        caption:
          type: string
        hook:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
        imagePrompt:
          type: string
        imageUrl:
          type: ["string", "null"]
          description: Base64 data URL of the AI-generated image
        platform:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - campaignId
        - day
        - caption
        - hook
        - cta
        - hashtags
        - imagePrompt
        - platform
        - createdAt
        - updatedAt
    UpdatePostBody:
      type: object
      properties:
        caption:
          type: string
        hook:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
        imagePrompt:
          type: string
        platform:
          type: string
    GeneratePostImageBody:
      type: object
      properties:
        customPrompt:
          type: string
          description: Custom image prompt from the user (overrides the post imagePrompt)
        size:
          type: string
          enum: ["256x256", "512x512", "1024x1024"]
          description: Image output size
        model:
          type: string
          enum: ["nano", "mini", "pro"]
          description: AI model quality for prompt enhancement
...